Import error for mode names (#1600)
* removed credentials

* import error solved
