+++ frontend/src/components/ChatBot/ChatInfoModal.tsx
@@ -23,12 +23,12 @@ import ChunkInfo from './ChunkInfo';
 import EntitiesInfo from './EntitiesInfo';
 import SourcesInfo from './SourcesInfo';
 import CommunitiesInfo from './CommunitiesInfo';
-import { chatModeLables, chatModeReadableLables, mergeNestedObjects, llms } from '../../utils/Constants';
+import { chatModeReadableLables, mergeNestedObjects, llms } from '../../utils/Constants';
 import { Relationship } from '@neo4j-nvl/base';
 import { getChatMetrics } from '../../services/GetRagasMetric';
 import MetricsTab from './MetricsTab';
 import { Stack } from '@mui/material';
-import { capitalizeWithUnderscore, getNodes } from '../../utils/Utils';
+import { capitalizeWithUnderscore, chatModeLables, getNodes } from '../../utils/Utils';
 import MultiModeMetrics from './MultiModeMetrics';
 import getAdditionalMetrics from '../../services/AdditionalMetrics';
 import { withVisibility } from '../../HOC/WithVisibility';
