+++ frontend/src/components/ChatBot/Chatbot.tsx
@@ -36,11 +36,11 @@ import { useFileContext } from '../../context/UsersFiles';
 import { useCredentials } from '../../context/UserCredentials';
 import clsx from 'clsx';
 import ReactMarkdown from 'react-markdown';
-import { buttonCaptions, chatModeLables } from '../../utils/Constants';
+import { buttonCaptions } from '../../utils/Constants';
 import useSpeechSynthesis from '../../hooks/useSpeech';
 import ButtonWithToolTip from '../UI/ButtonWithToolTip';
 import FallBackDialog from '../UI/FallBackDialog';
-import { downloadClickHandler, getDateTime, shouldShowTokenTracking } from '../../utils/Utils';
+import { chatModeLables, downloadClickHandler, getDateTime, shouldShowTokenTracking } from '../../utils/Utils';
 import ChatModesSwitch from './ChatModesSwitch';
 import CommonActions from './CommonChatActions';
 import Loader from '../../utils/Loader';
