+++ frontend/src/components/ChatBot/ChunkInfo.tsx
@@ -7,9 +7,8 @@ import youtubelogo from '../../assets/images/youtube.svg';
 import gcslogo from '../../assets/images/gcs.webp';
 import s3logo from '../../assets/images/s3logo.png';
 import ReactMarkdown from 'react-markdown';
-import { generateYouTubeLink, getLogo, isAllowedHost } from '../../utils/Utils';
+import { chatModeLables, generateYouTubeLink, getLogo, isAllowedHost } from '../../utils/Utils';
 import { ThemeWrapperContext } from '../../context/ThemeWrapper';
-import { chatModeLables } from '../../utils/Constants';
 import GraphViewModal from '../Graph/GraphViewModal';
 import { handleGraphNodeClick } from './chatInfo';
 import { IconButtonWithToolTip } from '../UI/IconButtonToolTip';
